Remove empty subject check for schedule send messages
Given the current implementation was reusing the OnSendWithEmtpySubject
event, this caused the message to be sent immediately instead of
scheduled when the user answered positively to the dialog.

To avoid putting in place a dedicated event we're removing this
till the feature is defined (not sure we do want a dialog here).

ET-3275