Fix 'also archive' flag not being considered on label as bottomsheet
Due to how AnimatedContent works, by not definining an explicit 'contentKey' the
state is used to define when the animation should be run.

Given in the labelAs case the state changes each time a label is checked (as well
as when the 'done' button is clicked), while the 'also archive' option is stored in
a mutable state on the view and only passed when user clicks 'done', the check of
a label was causing the animation to run again, causing the composable to leave
composition fully and re-enter it and the internal 'also archive' state to be reset.

The usage of a contentKey solves this by avoiding the animation while
the state is still `LabelAs.Data`.

ET-3543