Stop init and exit composer when opening existing draft fails
Reasons for failure opening an existing draft are not recoverable, such as the
draft not existing or the body not being available locally (eg. because we're
offline).
This is fixing a crash where opening a draft that's not available locally would
continue the composer init process, triggering the automatic save for subject,
recipients etc which in turn caused a crash on the data layer as the "Draft"
object (on rust side) did not exist (as init failed).

ET-3572