Adapt conversation items spacing to cover the bottom radius on expanded items
Due to the issue detailed in 9fdaee8
(google issue: https://issuetracker.google.com/issues/429387112)
the conversation item now has the rounded corners shape also
applied to the bottom. To avoid it being visible, we reduce
the padding that expanded conversations apply on the bottom
so that the negative margin used by lazycol to overlap the items
still overlaps enought to cover the corners.

ET-3593