Fix recomposition issue that was resetting signature when typing recipients after changing sender
Change sender action triggers a side effect that refreshes the display body of the message
(in order to update the signature to the new sender).
The "initial display body" state param, that's used when first opening the composer to show
the signature, was not being updated.
Given that when we add recipients the suggestions replace the body, effectively removing
the webview from composition, when going back from the suggestions the "initial display body"
was used once again to init the webview. In case the user didn't type anything after chaging the
sender, such initial state would still be referring to the previous sender's signature.

Updating the initial display body state as well when changing sender fixes this.

ET-3671