Drop logic to set hardcoded min height message body webview when empty
Drop as well the setting of any height at all through layout params.
This seems to solve the majority of the issues with expaning a message
or its quote. Allowing the webview to measure itself and just 'wrap content'
makes for a smooth animation and no blocked halfway loaded messages.

This will regress two other issues, at this stage:
- Opening message with empty body shows blank message
- (Potentially, didn't verify): scrolling quickly in large converasations might
have a non-smooth effect due to not catching webview height. I didn't see a big
difference in this regards.

ET-3517