Linkify urls when they are included in body, div or span tags
When sending HTML mime messages from certain clients the urls are
automatically linkified (ie. transformed to <a> tags) at
send time, while from other clients they are not.
Due to this, we need to be able to detect urls that are into html containers
and not already linkified. We do this with a positive lookback clause based
on a tags whitelist.

ET-3764