Bump rust to 0.99.3
This fixes the 'message with empty body not loading' (ET-3517)
and the 'message with single inline image as body not loading'
issues.
Given the first issue was caused by lack of padding in the root
element of the html, the padding was added back in the stylesheet by rust
and we removed padding from the webview on client side.

NOJIRA