Show remaining attempts when going back to the app while they are less than threshold
When entering the wrong pin 7 times (with the current max of 10 attempts)
the client shows an error letting the user know that they have 3,2,1
attempts remaining.
If the user kills the app while seeing this error and comes back to it,
the error is not shown till another incorrect attempt is performed,
which could mislead the user to belive the count is reset (while, it isn't).

With this commit the client shows the count also at app start, if less than 3
remaining.

ET-3809