Update message query implementation to adapt to new rust pagination API
The main difference of the new API is that the paginator methods (next, fetch)
do not return a list of items anymore but they just return "Ok/Error".
The items are exposed through an updated callback that is given to the paginator
at create time, and which lets clients know the 'type' of items returned (Append,
Refresh, None, Error...).
The callback can be fired without client actions as well, case in which, with the
exception of events that needs to trigger invalidation, we ignore it (as to display
items on the view, the request needs to be initiated by the client for paging3
to work).

- Adapt return type of message paginator classes to reflect rust API changes
  - The method to create a paginator (scrollMessagesForLabel) now takes a
ConversationScrollerLiveQueryCallback;
  - Return type of nextPage and reload methods are now either Error or Unit;

- Introduce a flow in RustMessageQueryImpl to receive items from the callback
and emit them "syncronously" though the getMessages method (as paging lib expects
items sync)

ET-3836