Update paging cache with invalidation filter to accept invalidation events without a fetcher
The new scroller API introduced in rust 0.100 changes the inner workings of
the callback, which is now only fired when one of the items being displayed
did actually change (vs when any table in the DB changed as before).
This includes handling for cases such as user opening a conversation
for which there are no messages locally - messages are loaded and written to DB
without causing invalidation of the conversation scroller.

Due to the above, clients do not need to check for data changes manually but
can trust the callback to perform an invalidation.

ET-3836