Change conversation paginator to invoke rust's forceRefresh method
This is needed as the refresh method was internally changed to return
the 'smallest possible update' of items in the list (using indexes).
This optimization was introduced for the iOS client which handles pagination
customly; It doesn't fit well in android where the current pagination impl relies
on google's paging3 which expects immutable data snapshots and disallows
granular updates.
The force refresh will return the whole data set refreshed as a 'ReploadFrom(idx=0)'

ET-3836