Update settings webview call to allow showing 'view model' toggle
In order for the setting webview to show the toggle, we need to fork the session
passing the product-name ("android-mail") as well as update each settings webview
url to pass a new app-version param.
The latter needs to be formatted as "android-version@7.0.3" without any suffix
(eg. -alpha+ab123cd) as that makes the request fail.

Due to this, usage of AppInformation class doesn't fit well and I went for
a new build config field with this purpose only (will be dropped when
native settings are introduced)

ET-3877