Mailbox refreshes when view mode preference is changed
This required to observe the 'view mode changed' as a trigger for re-creating
the paginator as well as adding a new method to RustMailboxFactory to allow
creating Mailbox skipping cache.

The usefulness of the cache is being evaluated in parallel. At now, we never hit
cache from the MailboxViewModel entry point, as for any given label we only
create the paginator once. The hits are instead in the entry points from message
and conversation data source / query which are used in the detail and for ops
related to the actions (get list of actions, execute action).

ET-3877