Fix mailbox loading getting stuck when switching user account
Due to the new 'observeViewMode' trigger (introduced to allow timely refresh on
the mailbox when view mode pref is changed) not re-triggering when re-subscribed
for a different userId with the same view mode as the previous, the paginator was
not being re-created resulting in the skeleton loading being show till the app
restarted.

Adding the userId as param to the 'current view mode' cache solves the issue.

ET-3877