Refresh UI when rust scroller emits a 'refresh all' signal without it being requested
The issue is that when moving items from locations that have few of them (or,
when moving the first item in any Mailbox), the item would be moved successfully
but the UI woulnd't be updated.

This issue happens when rust returns a ReplaceFrom(0) event, which contrarily to
any ReplaceBefore or a ReplaceFrom(>0)  does not trigger a "reload" on client
side (as "ReplaceFrom 0" is representing the "reloaded" signal itself).

Due to a paging3 constraint (ie. the UI doesn't react to pagination events unless
it did trigger a request first) this causes it not to "refresh" automatically in
this case.
It's mostly visible in "custom folders" as the ReplaceFrom(0) is triggered
in locations with fewer items, which often match custom ones.

The workaround consists in adding a flag to sync the refresh requests coming
from the UI with the ReplaceFrom signals coming from rust, so that we only returns
the items to be refreshed when UI did request it and we do trigger an invalidation
otherwise (which will cause the UI to request a refresh).

ET-3937