Update rust draft data source DI bindings to make it Singleton
This is a tradeoff to allow the Composer Screen to split functionality between
multiple view models.
Due to the setup of the rust library, objects (ie. the rust Draft) are held in memory
by clients. In this case, the "Draft data source" is the one holding an instance the the
rust Draft; The instance of draft data source is bound to ComposerViewModel and allows
to operate on the same "draft" without passing any identifiers (as by design,
there can only be one draft opened at a given time).

Having the Draft Data source bound to the viewModel scope gave us the guarantee that a fresh instance
would be injected for each composer session, ensuring usage of incorrect data would be avoided.
This also creates a big limitation as only one ViewModel (the composer's) can access the draft instance,
meaning any and all logic needs to go thorugh the composer VM.

This approach doens't have potential to scale well and already limited us in features such as
"schedule send" or "message body webview" which could have both had their own view model.

Thus, injecting as a signleton will allow to reuse the same instance anywhere (and adding new view models)
at the cost of having to "reset" the instance when the composer session closes.

ET-3945