Update Rust Draft Data Source to rely on new DraftCache implementation
The root reason for this new implementation is a mix of:
1. the willingess to hold the rust draft instance in memory in order to allow
breaking down composer's functionality in multuple VMs
2. the share via feature which creates the additional possibility of having two
ComposerVM instances alive at the same time

Point 1 (commits 4d43d4a48 and 1268129ee5) relies on the fact that only one
ComposerVM is in memory at any given time and thus the "draftwrapper" instance
cached in rust data source is only one, always valid and reset when the composer is closed.

Point 2 breaks that, as after share via composer is closed and "clears" the
cached draft, if user goes back to a previous, in app instance of composer
any operation will fail as the DraftDataSource has no cached (rust) draft anymore.

The solution consist in two parts:
- A list of "active composer instances" is held in memory
(Composer registers and un-registers to ActiveComposerRepo on init / clear)
- DraftCache caches the given DraftWrapper (retrieved at draft open / create op,
which is always the first operation that happen when opening a composer) using
the latest recorded "active composer" as a key.

This enables the cache to return the correct draft instance for the currently
active composer, and thus allow for two instances to live at the same time.

**Alternatives**
- Keying the Draft instance by messageId would have been clearer; This doesn't
work (in a straightforward fashion) as rust's "create" draft operation doesn't
immediately generate a `messageId` (which is only created when some content is
typed, which breaks this approach as at least one call need to happen without messageId).
- Keying by composer instance, taking it from the backstack instead of registering
/ unregistering manually: didn't find a reliable way to access backstack (as this
has to start from the data layer where the cache is and the composer NavController
which owns this backstack lives in the Home composable)

ET-3945