Chip list state exposes immutable list to onListChanged
This will help down the line if removing the de-duplication client side,
as doing that withouth passing the immutable list would break the addition
or recipeients as the call to chipListState.updateitems would be handling
it's own list and clearing it would cause state loss.

We're not dropping already the de-duplication due to the custom logic that's
run on the draftDataSource to define which recipients to update (which would
exclude a duplicated recipient from being added to rust in the first place,
making it not trivial to rely on rust for de-duplication).

ET-2160