Introduce ChipState.Validating
Since the validation is now on Rust side, we should append the chips in the "Validating" state to prevent race conditions where the user could potentially send a message to an invalid address mid-validation. There is currently no visual clue about the validation being in place, so that needs to be added.

ET-4712