Avoid repeated navigation to camera picker when back from message password screen
When adding a picture and then a message password, navigating back from the latter screen would
cause a navigation to picture picker to be re-triggered, due to the CameraPicturePicker
composable being re-composed and finding the previous state already set in the VM to the
FileInfo one which triggers navigation.

Resetting the camera picker state to 'Initial' after leaving the camera solves this.

ET-4752