Separate inline attachment deletion from the removal from the UI
With the previous solution, deletion of the inline attachment (eg. through tapping on it and
choosing 'remove' from the bottomsheet) would trigger a strip from body which would in turn trigger
a delete request (which would trigger a strip from body which would stop the chain as no image
would be there the second time around).
Stripping from body would in turn trigger deletion which triggers strip (and stops as no img).
Overall, it resulted in redundant events and errors and prevented the 'convert inline image to
standard' feature to be put in place (as it requires the ability to strip the image from UI without
deleting the attachment).

ET-2936