Stop init of observers when creation of rust draft fails
When a composer is opened, a corresponding 'Draft' object is created on rust
side (either a new draft, an existing one or a new one with a parent (ie. reply, fw)).
If the Draft creation operation fails, we display an unrecoverable error and close the
composer, but we also need to ensure we stop the initialization of futher watchers (eg.
attachments, recipients etc) otherwise they'll try to access the draft expecting it to be
there and fail when not finding it.

ET-5071