Adapt paginator implementation to new rust API
The main change being that 'unread filter' and 'show spam and trash' are not anymore
part of the 'scroller' constructor but are methods on its instance.
Consequently, clients must not re-create the whole scroller when these change but
just update the filtering thorugh the new method.
Scroller should still be re-created when the labelId or scroller type (search / default)
change.

NOJIRA