Fix moving a message to trash in the details screen doesn't have an immediate effect
The reason for this was that the external coroutineScope in
RustConversationDetailQueryImpl would get cancelled sometimes
when performing certain actions or even by opening bottom sheets.

The solution is to use the callbackFlow's internal scope instead
and bind the job's lifecycle to the flow itself.
That way external scope cancellations no longer affect it and
the flow automatically stays active for as long as it is needed.

ET-5197