Allow editing quoted html content when replying / forwarding
The "automatically collapse quote" feature introduced some css and js
updates that had a side effect of not allowing to edit the quote.

The fix consists in two parts:
- Do not move cursor away from the protonmail-quote element when the quote is
expanded (this check is needed to ensure the collapsed quote can't be selected
or deleted; when expanded, quoted html should be treated like any other text
in the composer and allow these ops);
- The css matcher that prevents quote deletion by making it not editable matched
any quote that didn't have the "data-expanded" attribute, which would also
match nested quotes.

ET-5226