Fix issue where 'include spam and trash' filter would remain visible after exiting search
The 'isExpandable' method which exposes whether a given location (label) should
display the 'include spam and trash' filter, was observed at the VM level
looking at labels changes.
Given the search feature doesn't change location, when a user entered search
the pager would be re-created and result in showing the filter.
When exiting search, this would trigger too early (due to change in state.observeSearchDataReady)
while the paginator didn't yet switch from the search to the default one.
This was causing the filter to be shown in locations that should not show it.

Moving the call to get 'is expandable' as a consequence of creating a new paginator
removed this race condition, allowing the API to only be called after the pager was created.

ET-5266