Fix issue where search results that are only in spam / trash don't show automatically when enabling the filter
when searching for a keyword for which the only matches are in spam or trash,
search results will be empty by default. If the user enables the "include spam / trash"
filter they should show right away, which they didn't due to the previous request still
being recorded as "pending" by our paginator logic.

This commit clears any pending request after the timeout for followups is
elapsed and the first response is returned.

ET-5335